Some Fixes …
This Commit contains a few Fixes and Changes for the Future of Suave7.

The general Icon-Template located in the Root contains a new
Overlay-Group, named "Colors", which makes it easier to create Icons
based on the Colors we all love. There’s also a new Color available :
Rosé Gold.

This Commit also contains additional Templates for WhatsApp and a few
new Icons.