Fixes, Fixes and Fixes again …
A few Years ago I’ve used Photoshop CS2 on my old Mac, which produced
some terrible PNG Previews. So I’ve fixed the Images, because I don’t
know if they cause any Problems on the iDevices. So now the Preview is
fixed and the Files are much smaller then before …

I’ve also changed the selected Layer in "Icons ( Aeneon, System )" and
"Icons ( Aeneon, Watch )", so it’s much easier to paste new Overlays in
it.

The last One is a small Addition for the Docks, with the Masks coming
from 7Eleven-Dock …