First Steps for a old iOS …
This Commit contains some new Templates and Icons for iOS 6.

So why iOS 6 ?

I own an iPad 2 since 2 Weeks and working on some private Stuff, but I
would like to share it with you. Currently on the iPad runs a very
basic Version of Suave7 ( or which I named SuavePad ). There’s a lot to
do, to theme all the Files coming from iOS 6, but it’s a first Step in
the right Direction …

I’ll upload the first Preview of the Package later this Month. This
Project can also be extended to support newer iPads with
Retina-Displays ( not my Goal, sorry ).

I hope you like it …