New iCreator-Icon …
Sometimes it's Time to move forward. The old Repo-Icon was created in 2016 and  during this Time a lot of Changes were made. my Repo isn't only for Suave7, so it was Time to reflect all the other Projects with a new Icon. Thanks to iOS 13, the main Part of the Icon is a Memoji, with minor Changes and a modern looking dark Background with Suave's Shape …